refractor: split TemplateDetail into multiple files
