+++ public/pages/CreateIndexTemplate/components/DefineTemplate/index.ts
@@ -1,0 +1,3 @@
+import DefineTemplate from "./DefineTemplate";
+
+export default DefineTemplate;
