+++ public/pages/CreateIndexTemplate/components/IndexSettings/index.ts
@@ -1,0 +1,3 @@
+import IndexSettings from "./IndexSettings";
+
+export default IndexSettings;
