+++ public/pages/CreateIndexTemplate/containers/IndexAlias/index.ts
@@ -1,0 +1,3 @@
+import IndexAlias from "./IndexAlias";
+
+export default IndexAlias;
