+++ public/pages/CreateIndexTemplate/containers/TemplateMappings/index.ts
@@ -1,0 +1,3 @@
+import TemplateMappings from "./TemplateMappings";
+
+export default TemplateMappings;
