+++ public/pages/CreateIndexTemplate/hooks.tsx
@@ -1,0 +1,10 @@
+import { EuiFormRowProps } from "@elastic/eui";
+import { FieldInstance, transformNameToString } from "../../lib/field";
+
+export const getCommonFormRowProps = (name: string | string[], field: FieldInstance): Partial<EuiFormRowProps> => {
+  return {
+    isInvalid: !!field.getError(name),
+    error: field.getError(name),
+    "data-test-subj": `form-row-${transformNameToString(name)}`,
+  };
+};
